Loader.load: fix dealing with non-string data type in FK/m2m columns
If the accession values found in FK/m2m columns are not stored as
strings in the DB we must convert them to the right type when we match
them up with existing rows.  This will be needed to support e.g. NCBI
taxonomy, where references to nodes use their number (taxid) which will
be stored as integer.