Make GLAMR frontpage resilient against DB connection failure
The idea is to allow the front page to display even if the DB is not
available.  Any DB access must be caught for this, and they may be
issues from strange places.  So the patch to BaxeMixin in this commit is
to make the base context processor immune to DB failure.