glamr/view: speed-up map point generation
Surprisingly, django.urls.reverse is quite slow, so generating thousands
of URLs for all those points on the map, becomes quite noticible.
Speeding this up turned out to be a bit invasive, so the new
fast_reverse function should be used with some care.

Anyways we see a 0.3 second improvement on the current frontpage.