glamr/tables: Fix: don't offer sorting for single column related table
Disable sorting.  The order_by field defaults to the column name
'object' which is not the right field.  To make sorting work the table
class with the column would need to know the involved specifics the the
relational field, so we're looking at generating the class via a factory
on-the-fly.  # Please enter the commit message for your changes. Lines
starting