glamr/views: keep sample-based dataset filters when referring back to samples
When filtering datasets based on some sample property, e.g. sample type,
we now consistently hold on to that filter when refering to the samples
of those selected dataset.  Previously, in some cases, all the datasets
samples would be related to.

  * adds reverse option to Dataset.shift()
  * add view attribute to Table base class in mibios.glamr.tables
  * preserve dataset filter in maps
  * preserve filter for related samples in dataset table

A follow-up commit will use these new features when dealing with
multiple samples at the coordinates on the map.