omics: Make update_analysis_status() work again
The imput file, imported_samples.tsv, changed format a bit, there are
now no
distinct lines for forward and reverse directions.

The method is now moved to Sample.loader.update_analysis_status() so it
is not affected by private samples (as it would be the case with
glamr's Sample.objects).  This required some shuffle of e.g.
MetaDataLoader from glamr.load to umrad.managers.