umrad/utils: ProgressPrinter: fix rate calculation for slow iteration
This patch changes when progress is printed.  Previously progress was
printed at each ring of the timer.  When dealing with large increments
but slow iteration, e.g. batch updates with postgress, a 20,000 row
update might take multiple seconds, the rate display would switch
between 0 and 20,000 per second.  With this commit, the timer's ring
will only ever set a checkpoint flag.  At each iteration step/increment
this flag is checked and, if set, progress gets printed.  This way the
true rate between consequent progress printing events is displayed.  The
disadvantage is that the timer interval is only the minimum interval
between progress printing events.  With slow iteraton the true interval
can be arbitrary long.  But the only true disadvantage is that the
estimated time remaining will not be updated.