omics/utils/Timestamper: fix logic when to save previous timestamp
The deleted conditional code was useless (I think.)  In any case, now
the previous timestamp is only ever updated at a newline.  This seems to
fix the problem when a text is added to a line and the full timestamp is
replaced by a delta.

We added some test code, namely printing a line longer than the terminal
is wide.  Adding more text to the line seems to print something similar
to two copies of the line instead, or so it seems on the narrow
terminal.  Dragging the terminal window border with the mouse to make it
wider, however, reveals that it is a single line (the line gets
unwrapped.)  In the file hard-copy it all look as it should.  I first
thought this problem was related to the above timestamp logic issue but
it seems these are two separate things after all.  Maybe the \r does not
work as expected on narrow terminals?  Anyways, that's why there is the
additional test code as part of the commit.