glamr/load: fix detecting timestamp precision
Switch the order in which we test, so now parse_date() is tried before
parse_datetime() as parse_date() will certainly fail if there is a time
given after the date.  But parse_datetime() will mostly accept dates
without time, just not dates with single-digit months (and maybe some
others).  Unsure why this used to work, maybe date formating changed a
lot in the input?