glamr+omics: check access to internal/test views at runtime
Previously these where enabled/disabled during urlconf setup, so on the
URL level.  This made is difficult to test these under different
settings.  Tests were adapted to check for enabling and disabling
settings as much as practicable.  Now, these URLs resolve but the view
might return a 404 depending on the settings.  The admin interface is
still setup early in glamr.admin.