glamr/search: Fix infinite loop processing spelling suggestions
In rare cases, e.g. underpopulated unique word table (when running
tests) a very short list of suggested spellings broke the limiting
algorithm.

Rewrote the suggestions processing to improve readability.  Though the
prove that the while loop will terminate is not immediate.

For a phrase of multiple misspelled words, the current limiting/picking
algorithm does not give as much weight to low distance suggestion as it
maybe should.