docker: mitigate crash on memcached failure
This adds the underdocumented 'noreply' option.  Large (>1 MB) responses
are not cached, using memcached default options, and it seems attempting
so, e.g. searching samples for "lake erie", raises an uncaught
exception:

pymemcache.exceptions.MemcacheServerError: b'object too large for cache'

leading to a 500 server error.  With the noreply option no exception is
raised and the page is computed like for a normal cache miss.