glamr/queryset: simplify full-text search query
Do the join with ContentType in python with a class-level cached map
from ContentType PKs to the model classes.  It's not clear if this saves
any time, but the postgres' query plan with EXPLAIN now looks much
simpler.

This also drops the order_by() on fields which had no use currently.