glamr: add globus file manager links to file tables
With this we have a working prototype for issue #35

Adds and utilizes an "extra_navigation" "framework" which allows a
table to pass table/model specific text/url pairs to the template which
are then displayed above the table as hrefs, similar to the existing
download links.  Maybe we can use the new framework for those download
links too at some point.