glamr/search: Re-designing secondary full-text search
As before do a fallback search if the primary search yields no hits.
But instead of fidgeting with mispelled words (which may not actually be
misspelled) we retrieve the tsquery string from postgresql and transform
it to give us a less restricted query (turning conjunctions into
disjunctions).

The result is much more robust in case of multi-word searches containing
a thing that did not make it into the word list but can still be found
via full-text search.