omics/models: Make migration independent of settings
The way PathField was designed made the path/prefix settings appear in
the migrations.  So path is now set via a method and the prefix validator got
taken out of the field and the corresponding checks get called from the
custom full_clean().  Also previously the validator was listed three
times in the migration (fixed now.)

This changes the previous omics/0016 migration, also carriying some
changes to SampleTracking from the previous commit.