omics/models: remove several status flag fields from Sample model and related clean-up
Removes those flags, and all their usage, that were taken over by the
new tracking facility.  Also removes flags that were not really in use.
For those there is some usage (in mostly dead) code remaining.  This
should be cleaned up later when we'll be wokring on binning.

Three flags (contig_abundance_loaded, contig_lca_loaded,
gene_alignment_loaded) remain for the time because related data is
currently present in the working databases but it's not yet clear if
we'll keep it.