umrad/utils: fix compile_ranges sometimes crashing or dropping a single
Dropping the element would have occurred on the for-break, so it's now
preceded by appending the element in the loop variable.  This would
only have happened under the sqlite DB backend but not under postgres so
had no impact on production data.

Check that single is not empty before removing END (avoids crash.)

Further, to improve readability, did some code shuffling to make it more
obvious that the loop is always entered at least once (for the END.)