glamr/models: Add data migration to type-changing migration 0022
This is quite a hack.  We add RunPython ops before and after the
field/column type changes.  The "before" op reads all data values and
then updates all of them to "0" which the DB can readily convert to
numeric or integer.  The "after" op then restores all the original
values that pass validation.  Empty string is restored as None/NULL.