umrad/utils/spec:  make loader spec more robust and flexible
This re-writes part of the spec-line parsing code to allow a prepfunc to
be specified but no field.  Absence of a field or prepfunc can be
explicit with a None, but that's not required now.  Errors messages
while parsing the in the spec should also be improved now.

A field in spec.fields can now be None, as such any users must be able
to handle that case, e.g. the main loop in Loader._load() now does a
continue on a None field just after running the prepfunc.

It is now possible for both field and prepfunc to be None for a column.
In this case the column is silently skipped.