Introduce the variable base template name
Adds the BASE_TEMPLATE_NAME setting, to be set to the name of the base
template of the site.  The mibios.context_processors.base context
processor needs to be added to the TEMPLATES configuration and will
populate all contexts with base_template_name.  In particular the mibios error
views (for glamr only the missing 403) depend on this.

Further, the model graphs and the omics sample tracking page make use of this.