glamr/load: set access groups when loading datasets
Since this sets an m2m relation between Dataset and Group (a
non-mibios/umrad model) we need some extra machinery (ugly hack?) to
tell the load() method that Group's accession field is Group.name and
how to filter for the name.  So this commit also adds the optional extra
kwargs to the InputSpec constructor which lets Dataset.loader pass that
Group.name info and then Loader.load() pick it up when processing the
m2m data.