glamr/search: fix transforming query for fall-back search
For certain queries running SELECT websearch_to_tsquery('english', 'a')
returns an empty tsquery that then caused a crash the parser
(Searchable.objects.parse_tsquery() to trip and crash the site with
DEBUG=True though in production it showed the 'no results' page.

This commit makes the whole flow more robust and documents return types
for a few methods.