glamr/models: Add access ArrayField to Dataset and Sample
These columns contain the Group primary keys, sorted, for those groups
that may access a sample or dataset.  Since 0 is never used as primary
key, it indicates that the record is public.

Queries to the field (for the purpose of establishing if access should
be given) should use e.g. access__overlap=[0, 1, 4] so always include 0.

ArrayField is postgres-only.  However, with default=None (instead of the
more apropriate default=list, additionally necessitating null=True) the
migration will succeed with Sqlite3.  The table should never be used
with Sqlite.

update_access() queryset methods are provided to sync/populate the
columns. For now, we keep Dataset.restricted_to as source of truth for
access data and as a fall-back for sqlite.

A follow-up commit will include maing use of the new columns.