glamr/search: fix crash on full text search coming from certain pages
Some, e.g. FilteredListView /filter/uniref100/, model-based views with
SearchFormMixin where the model is not one of the blessed searchable
ones, so e.g. UniRef100, that non-searchable model was passed to the
search_result view giving a 404 since invalid model.

The fix puts the "global search" placeholder as the model parameter in
those cases.