omics/managers: ensure the log file for update_from_pipeline() is closed
Saw a ResourceWarning issued inside the atomic_dry wrapper, so let's
make sure the file is written to and closed properly before it returns
to the atomic_dry wrapper.  Implemented via ExitStack since the log file
opening is conditional.  The input file is now also opened via the
ExitStack to keep the indentation level for the input file handling
portion as-is.