glamr/models: Fix regression, saving Dataset and Sample objects on Sqlite
The postgresql-only access (ArrayField) don't work with Sqlite.  This
commit allows us to do the initial INSERT when populating the database.
Some other thing may not work either, as we can save any non-NULL
values, but this is still enough for some dev/testing work.  In the
future, with Django 5.0 we may be able to declare the fields to use the
DB-level default.