umrad/manager: also pass field to Loader prep functions
This changes the calling conventions for all the prep methods: The calls
are now fn(value, field=field, obj=obj) and the keyword arguments are
called the "context."  As a convention, a method that does not need the
context is defined as "def meth(self, value, **ctx):". If the object is
used then we declare "def meth(self, value, obj, **ctx):" and similar
for the field or if both field and object are needed.  In any case the
ctx variable is ignored in the method body.  In the future if some prep
methods need more context, more keyword arguments can be added to the
calls without requiring any changes to the declarations of methods that
do not use the additional context.