umrad/models: tentative removal of FunctionName's FTS relation
Reverses part of ceb6c93e73274 which created a dependency to the glamr
app in umrad.  This should eventually be fixed by moving basic FTS
search into the mibios package.  But this is not the right time for it
as we're trying to get the hamb app started.

This will break the filtering on glamr's sample readabundance pages but
is needed to get the hamb app to work right now.