glamr/views: Add a sub-class view dispatcher to TableView
TableView gets a dispatcher to be used as view function in urlconfs.
This allows for model-specific subclasses of TableView (which are
expected to show up soon) to serve a generic URL with the model's name
as parameter.  TableView will build and keep a cache of the view
functions the first time the view dispatch is called.

Remove redundant FilteredListView class.

Correct doc string of ModelTableMixin.