glamr/views: Add view-subclass dispatcher to object-related list views
This extends the TableView subclass dispatching to object-related
listings.  The ToManyListView is replaced by an ObjectRelatedMixin that
is to be used together with TableView.  The dispatch determines the
model of the listing and then uses the appropriate TableView sub-class,
if one exist (defaulting to the generic TableView) to generate a view
also inheriting ObjectRelatedMixin.

The existing SampleListview is integrated, uses the new mixin, but will
not yet be picked up by the dispatcher, a follow-up will address this.