Split the Sample model
This splits the Sample model into:

  * Sample -- as in the biological specimen, with no regards as to any
    sequencing -- living mostly in the glamr app, but with a stub in
    omics, it keeps all the "meta" data, environmental, chemical data
    points etc.
  * SeqSample -- representing a sequencing run -- attached to Sample via
    parent FK relation.

This commit contains the automatically generated migration.  Those are
not meant to be applied as-is.  A follow-up will change them to migrate
data too.

Commented out any SeqSample<-Dataset relation.  Django's auto-migration
detection tripped on something there, always detecting some weird
alteration of the omics.Sample's dataset field.  Not sure yet if it'll
get re-introduced, do we need Dataset in the omics app at all?

Lots of regression are expected.  Data loading will be addressed with a
follow-up.

The first major commit to address issue #72.