omics: sample model split second migration
Migrate several FK relations from using Sample to SeqSample.  This is
the first step, adding a new nullable FK to SeqSample and populating it.
Assumes there is still a 1-1 relation between Sample and SeqSample.

In a follow-up migration the old FKs to Sample are removed.

WIP: still need to write the migrations