glamr: Add dump_test_data and load_test_data commands
These two commands can be used to generate smallish test databases with
postgresql database backends.  The dump command takes sample IDs as
arguments.  The idea is to pick a few samples with relatively low read
or contig counts and generate a small but fully functional and
representative test database in about five to ten minutes.