glamr/querysets: add more biosample deduplication tooling
The new dedup_biosample_migrate() method does the actual deduplication/
It will deduplicate those biosamples for which the operation is possible
without data loss.  There is no facility for automatic reversal.  Manual
reversal is possible via copying a biosample objects and changing the
sample_id to correspond to the one used in the seqsample and changing
the m2m relation back appropriately.